test.py
try:
    from langchain_classic.chains import create_retrieval_chain
    from langchain_classic.chains.combine_documents import create_stuff_documents_chain
    from langchain_community.document_loaders import DirectoryLoader, TextLoader
    from langchain_text_splitters import RecursiveCharacterTextSplitter
    import streamlit as st
    import os
    from dotenv import load_dotenv
    from langchain_groq import ChatGroq
    from langchain_core.prompts import ChatPromptTemplate
    from langchain_community.chat_models import ChatOllama
    from src.rag_engine import get_rag_chain
    from src.ingestion import load_documents, chunk_documents
    from src.vector_store import create_vector_db
    print('import successful')
except Exception as e:
    print('import failed', e)